front_end/requirements.txt
streamlit>=1.28.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
haven>=0.1.0
